+++ lib/features/review_your_choosen_meals/presentation/review_your_choosen_meals_screen.dart
@@ -93,7 +93,7 @@ class ReviewYourChoosenMealsScreen extends StatelessWidget {
                         log("Button Taped : Remove");
                         showSwapMealBottomSheet();
                       },
-                      kcalValue: data.kcalValue,
+                      kcalValue: data.kcalValue.toInt(),
                       mealType: data.mealType,
                       mealTitle: data.mealTitle,
